api/requirements.txt
fastapi
uvicorn[standard]
requests
redis
rq
python-dotenv
aiofiles